% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidMarkdown-package}
\alias{AcidMarkdown}
\alias{AcidMarkdown-package}
\title{AcidMarkdown}
\description{
Toolkit for extending the functionality of R Markdown.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/acidmarkdown/}
  \item \url{https://github.com/acidgenomics/r-acidmarkdown/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-acidmarkdown/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
