% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/markdownPlots-methods.R
\name{markdownPlots}
\alias{markdownPlots}
\alias{markdownPlots,list-method}
\title{Markdown plots}
\usage{
markdownPlots(object, ...)

\S4method{markdownPlots}{list}(object, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{...}{Additional arguments.}
}
\value{
Graphical output of plots.
Separated by Markdown headers.
}
\description{
Markdown plots
}
\details{
Supports using a named \code{list} containing multiple \code{ggplot} objects, which
can be used in an R Markdown report, separated by headers. Internally, the
headers are generated with the \code{markdownHeader} function.
}
\note{
Updated 2022-05-24.
}
\examples{
## Base R plotting example.
data(airquality, package = "datasets")
p1 <- hist(airquality[["Ozone"]], plot = FALSE)
p2 <- hist(airquality[["Wind"]], plot = FALSE)
list <- list("ozone" = p1, "wind" = p2)
markdownPlots(list)
}
\seealso{
Examples taken from:
https://bookdown.org/rdpeng/exdata/the-base-plotting-system-1.html
}
