% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/markdownTables-methods.R
\name{markdownTables}
\alias{markdownTables}
\alias{markdownTables,list-method}
\title{Markdown tables}
\usage{
markdownTables(object, ...)

\S4method{markdownTables}{list}(object, captions = NULL, force = FALSE)
}
\arguments{
\item{object}{Object.}

\item{captions}{\code{character} or \code{NULL}.
Captions.
If \code{NULL}, the names of the primary input list will be used automatically.}

\item{force}{\code{logical(1)}.
Force knit output using \code{knitr::asis_output()}.
\emph{Recommended for development and unit testing only.}}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item knit call: \code{asis_output}
\item Otherwise: \code{list}.
}
}
\description{
Markdown tables
}
\details{
\code{knitr::kable()} now supports multiple tables as a \code{list} for the \code{x}
argument, but it still only supports a single caption. \code{markdownTables}
extends this functionality, but supporting captions for each table.
}
\note{
Updated 2022-05-24.
}
\examples{
list <- list(
    "iris" = head(datasets::iris),
    "mtcars" = head(datasets::mtcars)
)
captions <- c(
    "iris" = "Edgar Anderson's iris data",
    "mtcars" = "Motor Trend car road tests"
)
markdownTables(list, captions = captions)
}
\seealso{
\itemize{
\item \code{knitr::kable()}.
\item \href{https://stackoverflow.com/a/35149103/3911732}{Stack Overflow post}.
}
}
