% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownPlots.R
\name{markdownPlots}
\alias{markdownPlots}
\alias{mdPlots}
\title{Multiple Markdown plots}
\usage{
markdownPlots(list, headerLevel = 2L)
}
\arguments{
\item{list}{\code{list}.
Named list containing \code{ggplot} objects.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}
}
\value{
Graphical output of plots.
Separated by Markdown headers.
}
\description{
Supports using a named \code{list} containing multiple \code{ggplot} objects, which
can be used in an R Markdown report, separated by headers. Internally, the
headers are generated with the \code{markdownHeader} function.
}
\note{
Updated 2021-02-04.
}
\examples{
## Base R plotting example.
data(airquality, package = "datasets")
p1 <- hist(airquality[["Ozone"]], plot = FALSE)
p2 <- hist(airquality[["Wind"]], plot = FALSE)
list <- list("ozone" = p1, "wind" = p2)
markdownPlots(list = list)
}
\seealso{
\itemize{
\item Examples taken from:
https://bookdown.org/rdpeng/exdata/the-base-plotting-system-1.html
\item \code{cowplot::plot_grid()}.
}
}
